feat(steps): add install-node step
This step allows installing a defined version of node via nvm.
This version can be passed to agents via the `NX_CLOUD_NODE_VERSION`
environment variable or via populating a `.nvmrc` file in the workspace.
```yaml
- name: Install Node v18
  uses:
  'nrwl/nx-cloud-workflows/<version>/workflow-steps/install-nodes/main.yaml'
  env:
    NX_CLOUD_NODE_VERSION: 18

```

if a `.nvmrc` file exists in the workspace (via a checkout step), then
  the `NX_CLOUD_NODE_VERSION` can be omitted.

```yaml
- name: Install Node v18
  uses:
  'nrwl/nx-cloud-workflows/<version>/workflow-steps/install-nodes/main.yaml'
```