feat(install-node-modules): allow controlling legacy-peer-deps for npm
Most people should not being using legacy peer deps anymore, but we
understand the compat needs and could be a breaking change to just
change this from under people.

This PR introduces a new input to the install-node-modules step to
enable/disable the usage of `--legacy-peer-deps` arg for `npm ci`
install command.

Since it's a lot of overhead to ask people to create a launch template
just to control 1 input, this also introduces the
`NX_CLOUD_NPM_LEGACY_INSTALL` env var that can be sent via the main
agent to controll this behavior as well.