fix(launch-template): remove playwright from browser cache
Playwright recommends not caching the binaries since restoring from the
cache usually takes longer than just downloading from the CDN.

> Caching browser binaries is not recommended, since the amount of time
it takes to restore the cache is comparable to the time it takes to
download the binaries. Especially under Linux, operating system
dependencies need to be installed, which are not cacheable.

https://playwright.dev/docs/ci#caching-browsers